refactor: move v4 processors to reactor plugin LogRequestProcessor and LogResponseProcessor are not moved with this commit because "v2" code relies on it. it will be done in a separated commit
