+++ gravitee-apim-plugin/gravitee-apim-plugin-reactor/gravitee-apim-plugin-reactor-v4/src/main/java/io/gravitee/apim/plugin/reactor/processor/reporter/EventNativeReporterProcessor.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.gateway.reactive.handlers.api.v4.processor.reporter;
+package io.gravitee.apim.plugin.reactor.processor.reporter;
 
 import io.gravitee.gateway.reactive.core.context.MutableExecutionContext;
 import io.gravitee.gateway.reactive.core.processor.Processor;
@@ -21,8 +21,6 @@ import io.gravitee.gateway.report.ReporterService;
 import io.gravitee.reporter.api.v4.metric.Metrics;
 import io.reactivex.rxjava3.core.Completable;
 import lombok.extern.slf4j.Slf4j;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * @author David BRASSELY (david.brassely at graviteesource.com)
